Add a class called IdentityMatrix
We now use `Matrix<int>(shape)` to create a matrix whose elements are
all zero. Now `Matrix<int>(shape, IdentityMatrix())` will create an
identity matrix.

See #87.
