+++ src/include/identity_matrix.h
@@ -1,0 +1,8 @@
+#ifndef MCA_IDENTITY_MATRIX_H
+#define MCA_IDENTITY_MATRIX_H
+namespace mca {
+struct IdentityMatrix {
+    explicit IdentityMatrix() = default;
+};
+}  // namespace mca
+#endif
