+++ src/include/matrix_declaration.h
@@ -1,8 +1,6 @@
 #ifndef MATRIX_DECLARATION_H
 #define MATRIX_DECLARATION_H
 namespace mca {
-struct Shape;
-
 template <class ELEMENT_TYPE = double>
 class Matrix;
 }  // namespace mca
